fix: 0.2.0 M1 — 问题档修复 + issue#3/#4 + 退出警告清理
- get_component 未挂载返回 null；新增 entity(id)/has_entity(id)
- StringFixed 超长截断保留整字符并 push_warning；count==0 安全存空串
- ComponentRegistry 改堆分配单例(永不 delete)，name_of 改按值返回；
  反初始化先 shutdown() 再 clear()
- 新增 World::reset + ObserverDispatch::clear + VECSWorld::shutdown
- Observer 字段级订阅 set_fields/get_fields + change_tick 节流 set_throttle_tick
- issue#3: change_tick_ 与版本号全部 uint32→uint64；generation 到
  UINT32_MAX 槽永久退役
- issue#4: serialized_component_size + validate_packet，apply 前先校验，
  截断/未知 schema/id 冲突包级拒绝，无部分状态
- 回归新增 T15-T20（entity lookup / stringfixed / observer 字段节流 /
  网络校验 / shutdown），T1-T20 共 57 断言全绿
- plugin.cfg author 改 ASDF Studio

+++ src/core/archetype.cpp
@@ -32,7 +32,7 @@ const Column &Archetype::column(ComponentTypeId p_t) const {
 	return columns[idx];
 }
 
-uint32_t Archetype::add_entity(Entity p_e, uint32_t p_change_tick) {
+uint32_t Archetype::add_entity(Entity p_e, uint64_t p_change_tick) {
 	uint32_t row = static_cast<uint32_t>(entities.size());
 	entities.push_back(p_e);
 	enabled.push_back(1);
